+++ packages/leavittbook/src/demos/titanium-chip-multi-select/titanium-chip-multi-select-playground.ts
@@ -8,6 +8,9 @@ import { TitaniumChipMultiSelect } from '@leavittsoftware/web/titanium/chip-mult
 import '@material/web/chips/input-chip';
 import '@material/web/icon/icon';
 import '@material/web/button/outlined-button';
+import '@material/web/button/filled-button';
+import '@material/web/button/filled-tonal-button';
+
 import '@leavittsoftware/web/titanium/chip-multi-select/chip-multi-select';
 
 const chipLabels = ['Dog', 'Cat', 'Lion', 'Hedgehog', 'Turtle', 'Monkey', 'Owl', 'Peacock', 'Pigeon', 'Spider', 'Tortoise', 'Zebra'];
@@ -35,10 +38,11 @@ export class TitaniumChipMultiSelectPlayground extends LitElement {
         border: 1px solid var(--md-sys-color-outline);
         padding: 24px;
         border-radius: 8px;
-        display: flex;
-        flex-direction: column;
+        display: grid;
         gap: 12px;
         margin: 24px 0 36px 0;
+        resize: both;
+        overflow: hidden;
       }
 
       button-container {
@@ -118,6 +122,45 @@ export class TitaniumChipMultiSelectPlayground extends LitElement {
             >Reset</md-outlined-button
           >
         </button-container>
+      </div>
+
+      <h1>Filled</h1>
+      <p>Example with filled variant</p>
+      <div>
+        <titanium-chip-multi-select
+          filled
+          demo2
+          label="Service Animals"
+          ?hasItems=${!!this.demoItems.length}
+          .supportingText=${this.supportingText ?? ''}
+          ?disabled=${this.disabled}
+          required
+        >
+          <md-filled-tonal-button
+            ?disabled=${this.disabled}
+            @click=${async () => {
+              this.demoItems.push(chipLabels[this.demoItems.length % chipLabels.length]);
+              this.requestUpdate('demoItems');
+            }}
+          >
+            Add Animal
+            <md-icon slot="icon">add</md-icon>
+          </md-filled-tonal-button>
+          ${repeat(
+            this.demoItems,
+            (o) => o,
+            (o, index) =>
+              html`<md-input-chip
+                label=${o}
+                closeable
+                ?disabled=${this.disabled}
+                @remove=${(e: Event) => {
+                  e.preventDefault();
+                  this.demoItems = this.demoItems.filter((_, i) => i !== index);
+                }}
+              ></md-input-chip>`
+          )}</titanium-chip-multi-select
+        >
       </div>
     `;
   }